chore(app-launcher): refactored to accommodate import app and create app workflows
In addition to the launcher being able to support application creation, it must also support the ability to import an existing application.
The step names may be similar, but the UXD mockups show very different step content.
Thus, we need to have multiple step components for the summary, progress, pipeline, and import existing app (i.e., the gitprovider).